Fix @smooth-ui/core-em TypeScript definitions
Adjust the build step so core-em and core-sc have their own independent
type definitions, and copy the shared definitions to a
'shared/index.d.ts' module.
